requirements.txt
openai~=1.63.2
opencv-python~=4.10.0.84
pillow~=11.0.0